mv fake fact dir to tools
